revert request class move
